implement dynamic API for likelihoods
