+++ src/mofaflex/_core/api/types.py
@@ -1,10 +1,27 @@
-from ..terms import TermWrapper
-from ..utils import building_docs
-from . import terms
+def _generate():
+    from types import ModuleType
 
-if building_docs():
-    for term in dir(terms):
-        globals()[term] = getattr(terms, term)
-else:
-    for term in dir(terms):
-        globals()[term] = TermWrapper
+    from ..likelihoods import Likelihood
+    from ..terms import Term
+    from ..utils import building_docs
+    from . import likelihoods, terms  # noqa: F401
+    from .utils import DynamicAPIWrapper
+
+    for docsapi, basecls, subclss in (
+        ("terms", Term, Term.known_terms()),
+        ("likelihoods", Likelihood, Likelihood.known_likelihoods()),
+    ):
+        mod = ModuleType(docsapi)
+        if building_docs():
+            apimod = locals()[docsapi]
+            for wrapper in dir(apimod):
+                setattr(mod, wrapper, getattr(apimod, wrapper))
+            setattr(mod, basecls.__name__, None)
+        else:
+            for clsname, subcls in subclss.items():
+                setattr(mod, clsname, DynamicAPIWrapper[subcls])
+            setattr(mod, basecls.__name__, DynamicAPIWrapper[basecls])
+        globals()[docsapi] = mod
+
+
+_generate()
