+++ src/mofaflex/_core/likelihoods/negativebinomial.py
@@ -1,4 +1,7 @@
+from typing import Literal
+
 import numpy as np
+import pandas as pd
 from numpy.typing import NDArray
 
 from .. import utils
@@ -103,6 +106,11 @@ class NegativeBinomial(Likelihood):
         data -= self._shift[group_name][feature_idx]
         return data
 
-    @property
-    def dispersion(self):
-        return self._dispersion
+    @Likelihood._api
+    def get_dispersion(self, moment: Literal["mean", "std"] = "mean") -> pd.Series:
+        """Get the dispersion vectors for each view.
+
+        Args:
+            moment: Which moment of the posterior distribution to return.
+        """
+        return pd.Series(getattr(self._dispersion, moment), index=self._feature_names)
