+++ src/mofaflex/_core/terms/__init__.py
@@ -1,3 +1,2 @@
-from .api import TermWrapper
 from .base import Term
 from .mofaflex import MofaFlex
